RIPENCC Expert Bot improvements and clean up
