+++ intelmq/bots/BOTS
@@ -215,6 +215,17 @@
                 "redis_cache_ttl": "86400"
             }
         },
+        "RIPENCC": {
+            "module": "intelmq.bots.experts.ripencc.ripencc",
+            "description": "RIPENCC is the bot resposible to get the correspondent abuse contact from source IP and destination IP of the events",
+            "parameters": {
+                "processing_interval": "0",
+                "redis_cache_host": "127.0.0.1",
+                "redis_cache_port": "6379",
+                "redis_cache_db": "5",
+                "redis_cache_ttl": "86400"
+            }
+        },
         "GeoIP": {
             "module": "intelmq.bots.experts.geoip.geoip",
             "description": "GeoIP (MaxMind) is the bot responsible to add geolocation information to the events (Country, City, Longitude, Latitude, etc..)",
