--- intelmq/bots/experts/ripeabuselookup/config-examples/BOTS
@@ -1,13 +1,0 @@
-{
- .....
-    "Expert": {
-        "ripeabuselookup": {
-            "module": "intelmq.bots.experts.ripeabuselookup.ripeabuselookup",
-            "description": "a bot which will check if the event falls into the RIPE region and if yes, will query stat.ripe.net for abuse contacts",
-            "parameters": {
-                "processing_interval": "0"
-            }
-        },
-    },
- .....
-}
