--- intelmq/bots/experts/ripeabuselookup/config-examples/runtime.conf
@@ -1,7 +1,0 @@
-{
- .....
-    "ripeabuselookup-expert": {
-        "processing_interval": 0
-    },
-  ......
-}
