--- intelmq/bots/experts/ripeabuselookup/config-examples/startup.conf
@@ -1,10 +1,0 @@
-{
- ....
-    "ripeabuselookup-expert": {
-        "group": "Expert",
-        "name": "ripeabuselookup",
-        "module": "intelmq.bots.experts.ripeabuselookup.ripeabuselookup",
-        "description": "a bot which will check if the event falls into the RIPE region and if yes, will query stat.ripe.net for abuse contacts"
-    },
- .....
-}
